Only grant a permission for a sub-tag if that permission is granted for its parent too
Previously if a parent tag was restricted but none of its sub-tags were, the sub-tags would be visible and the parent tag would be loaded in the tag list as a result. This adds permission logic so that sub-tags automatically assume the permissions of their parent as a minimum.

Fixes flarum/core#833