Safari iOS tag alignment patch
In Safari iOS, child tags are slightly higher than the parent tag.

I’m recommending adding this at the bottom of this file with the comment left in place so that it can be easily removed in the future as needed. 

Also, to be safe, using .mobile-safari class which is already automatically added to the #app class list when Safari iOS is detected. Thus, fix/patch will not affect other browsers.